requirements.txt
fastapi
uvicorn
litellm
pyyaml
python-dotenv
